ripple-backend-local/requirements.txt
# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1
httpx==0.27.2

# ── Config / validation ────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pyyaml==6.0.2

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib==1.7.4

# ── ML client (feature engineering for graph payload) ─────────────────────────
numpy==1.26.4

# ── Utils ──────────────────────────────────────────────────────────────────────
tenacity==9.0.0
loguru==0.7.2
aiohttp==3.11.7

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0